+++ tests/Views/Stringable.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Awesomite\ErrorDumper\Views;
+
+/**
+ * @internal
+ */
+class Stringable
+{
+    private $callable;
+
+    /**
+     * Stringable constructor.
+     * @param callable $callable
+     */
+    public function __construct($callable)
+    {
+        $this->callable = $callable;
+    }
+
+    public function __toString()
+    {
+        return call_user_func($this->callable);
+    }
+}